refactor: rename to TaskInfoField
